the free software manifesto
A spectre is haunting cyberspace - the spectre of free software. All the powers of the unfree Internet have entered into a holy alliance to exorcise this spectre: Apple and Microsoft, Facebook and Google, Oracle and legislation in their interests.

Where is the party in opposition that has not been decried as supporting free software by its opponents in power? Where is the opposition that has not hurled back the branding reproach of free software, against the more advanced opposition parties, as well as against its proprietary adversaries?

Two things result from this fact:  
1. Free software is already acknowledged by all closed-source powers to be itself a power.
2. It is high time that free software proponents should openly, in the face of the whole world, publish their views, their aims, their tendencies, and meet this nursery tale of the Spectre of free software with a manifesto of GNU itself.

To this end, free software activists of various nationalities have assembled on behalf of Richard Stallman and sketched the following manifesto, to be published in the C, Rust, JavaScript, Python, Java, TypeScript, Swift, Kotlin, Haskell and Go languages.